queue, cmd, ethereum: add tls
